Corrections for Pylint and PEP convention
